+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
+- Hide unwanted upgrades from site-creation and quickinstaller
+  [boulch]
+
 - Add local manager role and sharing permissions rolemap
   [boulch]